# swagger_poolhub.yaml
openapi: 3.0.3
info:
  title: PoolHub Entertainment Center Management API
  description: |-
    Đặc tả hệ thống RESTful API cho **PoolHub - Hệ thống quản lý trung tâm giải trí bi-a và dịch vụ** phục vụ môn học **PRN232 - Tuần 2**.
    
    Tài liệu này định nghĩa 100% tất cả các endpoint nghiệp vụ thực tế khớp với file thiết kế **Analysis PoolHub.xlsx**, hỗ trợ phân quyền truy cập chặt chẽ (Guest, Staff, Manager, Admin) và mô hình phản hồi JSON chuẩn hóa.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
paths:
...
  /sessions/start:
    post:
      tags:
        - Table & Session
      summary: Mở bàn chơi (Bắt đầu phiên chơi) (Staff, Manager, Admin)
      description: Khởi chạy một phiên hoạt động mới cho bàn trống khi khách vào chơi bi-a.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tableId
              properties:
                tableId:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Bắt đầu phiên chơi thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStartResponse'
        '400':
          description: Bàn bi-a đang được sử dụng hoặc đang trong trạng thái bảo trì
  /sessions/{sessionId}/end:
    post:
      tags:
        - Table & Session
      summary: Đóng bàn chơi & Tính tiền giờ (Staff, Manager, Admin)
      description: Kết thúc phiên chơi, tự động tính tổng phút hoạt động và quy đổi tiền giờ theo khung giờ vàng.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            example: ses_a1b2c3d4
      responses:
        '200':
          description: Kết thúc phiên chơi và chốt tiền giờ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEndResponse'
  /sessions/{sessionId}/switch:
    post:
      tags:
        - Table & Session
      summary: Chuyển đổi bàn chơi (Staff, Manager, Admin)
      description: Chuyển khách sang bàn chơi trống khác mà vẫn bảo lưu và tiếp nối thời gian tính giờ của phiên chơi cũ.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            example: ses_a1b2c3d4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetTableId
              properties:
                targetTableId:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Chuyển bàn thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Chuyển khách thành công từ Bàn 01 sang Bàn 02.

  # ==========================================
  # MODULE 3: BOOKING & ORDER
  # ==========================================
  /bookings:
    post:
      tags:
        - Booking & Order
      summary: Đặt bàn trước (Guest)
      description: Cho phép khách hàng đặt lịch hẹn giữ bàn trước qua ứng dụng Web (chọn ngày, giờ, loại bàn bi-a).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Đăng ký đặt lịch thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
    get:
      tags:
        - Booking & Order
      summary: Danh sách đặt lịch hệ thống (Staff, Manager)
      description: Xem toàn bộ danh sách đặt bàn của khách để thu ngân xác thực hoặc chuẩn bị bàn.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingResponse'
...
  /products:
    post:
      tags:
        - Booking & Order
      summary: Thêm sản phẩm/dịch vụ mới (Admin Only)
      description: Thêm đồ uống, đồ ăn nhẹ hoặc dịch vụ bổ sung vào thực đơn trung tâm bi-a.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Thêm sản phẩm thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    get:
      tags:
        - Booking & Order
      summary: Danh sách menu đồ uống & dịch vụ
      description: Lấy danh sách thực đơn đồ uống và dịch vụ bổ sung theo danh mục (hỗ trợ phân trang).
      responses:
        '200':
          description: Tải thực đơn thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductResponse'
  /products/{id}:
    put:
      tags:
        - Booking & Order
      summary: Cập nhật sản phẩm trong menu (Admin Only)
      description: Sửa tên sản phẩm, hình ảnh, đơn giá, số lượng tồn kho.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    delete:
      tags:
        - Booking & Order
      summary: Xóa sản phẩm khỏi thực đơn (Admin Only)
      description: Gỡ sản phẩm khỏi thực đơn bi-a.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 101
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Sản phẩm đã được gỡ bỏ khỏi thực đơn.
...
  /invoices:
    post:
      tags:
        - Invoice & Payment
      summary: Tạo hóa đơn thanh toán khi đóng phiên (Staff, Manager, Admin)
      description: Hệ thống tự chốt phiên chơi, tính tổng tiền giờ + nước để tạo bản ghi hóa đơn nháp.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
              properties:
                sessionId:
                  type: string
                  example: ses_a1b2c3d4
      responses:
        '201':
          description: Tạo hóa đơn nháp thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      invoiceId:
                        type: string
                        example: inv_v9u8t7s6
                      sessionId:
                        type: string
                        example: ses_a1b2c3d4
                      totalPayable:
                        type: number
                        example: 330000.0
  /invoices/{invoiceId}:
    get:
      tags:
        - Invoice & Payment
      summary: Xem chi tiết một hóa đơn lưu trữ
      description: Xem lại hóa đơn cũ đã thanh toán cùng đầy đủ tiền giờ, danh sách đồ uống gọi đi kèm.
      security:
        - BearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            example: inv_v9u8t7s6
      responses:
        '200':
          description: Tải hóa đơn thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetailResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SessionStartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Mở bàn thành công.
        data:
          type: object
          properties:
            sessionId:
              type: string
              example: ses_a1b2c3d4
            tableId:
              type: integer
              example: 1
            startTime:
              type: string
              example: "2026-05-24T17:50:00Z"
            status:
              type: string
              example: Active
    SessionEndResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            sessionId:
              type: string
              example: ses_a1b2c3d4
            totalMinutes:
              type: integer
              example: 90
            playTimeCost:
              type: number
              example: 180000.0
            status:
              type: string
              example: Ended
    BookingRequest:
      type: object
      required:
        - customerName
        - phone
        - bookingDate
        - bookingTime
        - tableType
      properties:
        customerName:
          type: string
          example: Tran Van B
        phone:
          type: string
          example: "0912345678"
        bookingDate:
          type: string
          example: "2026-05-25"
        bookingTime:
          type: string
          example: "19:30:00"
        tableType:
          type: string
          example: 9-Ball
    BookingResponse:
      type: object
      properties:
        bookingId:
          type: string
          example: bk_e5f6g7h8
        customerName:
          type: string
          example: Tran Van B
        phone:
          type: string
          example: "0912345678"
        bookingDateTime:
          type: string
          example: "2026-05-25T19:30:00Z"
        tableType:
          type: string
          example: 9-Ball
        status:
          type: string
          example: Pending
    ProductRequest:
      type: object
      required:
        - productName
        - category
        - price
      properties:
        productName:
          type: string
          example: Cà Phê Sữa Đá
        category:
          type: string
          example: Cafe
        price:
          type: number
          example: 29000.0
        stockQuantity:
          type: integer
          example: 50
    ProductResponse:
      type: object
      properties:
        productId:
          type: integer
          example: 102
        productName:
          type: string
          example: Cà Phê Sữa Đá
        category:
          type: string
          example: Cafe
        price:
          type: number
          example: 29000.0
        stockQuantity:
          type: integer
          example: 50
...
    InvoiceDetailResponse:
      type: object
      properties:
        invoiceId:
          type: string
          example: inv_v9u8t7s6
        tableName:
          type: string
          example: Bàn 02
        playTimeCost:
          type: number
          example: 180000.0
        totalOrderCost:
          type: number
          example: 135000.0
        discountAmount:
          type: number
          example: 15000.0
        taxAmount:
          type: number
          example: 30000.0
        totalPayable:
          type: number
          example: 330000.0
        paymentStatus:
          type: string
          example: Paid
...